numparam, change np_trunc() and np_round() return type to double

+++ src/frontend/numparam/general.h
@@ -47,8 +47,8 @@ char rc(void);
 
 int freadstr(FILE *f, SPICE_DSTRINGPTR dstr_p);
 
-long np_round(double d);        // sjb to avoid clash with round() in math.h
-long np_trunc(double x);        // sjb to avoid clash with trunc() in math.h
+double np_round(double d);        // sjb to avoid clash with round() in math.h
+double np_trunc(double x);        // sjb to avoid clash with trunc() in math.h
 double absf(double x); /* abs */
 long absi(long i);
